Break LocalDevice into own files.
